+++ packages/react/src/richtext/fixtures/code-component.tsx
